The line on the stars in google chrome appears because of the link before. I have tried to remove this several ways, setting for the html elements that encapsulate the rating the following style settings: text-decoration:none; border: none;
Then i tried to work on the div's properties but this didn't work either on chrome.

The final solution was to remove the div and the line dissapeared. The only problem now is that I can't align the <ul> to the right, neither encapsulate it in a span for example, in order to keep the alignment  and not to have the problem anymore.