Remix view - I still have to write the part where I send the added sound to the of sources (this worked in the original version but now I want to send the template). The part where the sound is added to the database is working.
In the urls I put the links for the remix view and the ajax solr search.

You can find the 2 views for these in the sounds and search. 

The templates can be found in the sound directory. 

The search is done through get (though at the beggining I was using the api) and the adding and removing with ajax post. 

Div removing and adding is handled with jquery.