Use denormalised counts on profile for number of sounds and posts
This query is significantly faster than performing a count on the
sounds table, however it may be out of sync based on #560, so needs
to be looked at.
We don't have a similar field for comments, so they still perform a
count, but we could reduce this page load by another 60-100ms by
adding such a field.