Updated sound model with 'processing_ongoing_state' field
Current 'processing_state' field will from now on only represent states 'PE', 'FA', and 'OK'. 'PE' (pending) is the default state the very first time a sound is processed and can not yet be moderated and published. The states 'OK' and 'FA' are used to represent that a sound has been successfuly processed or failed processing (respetively). We added the 'processing_ongoing_state' which will represent states 'QU' (queued), 'PR' (processing) and 'FI' (processing finished) states and can be used in combination with processing_state.

https://github.com/MTG/freesound/issues/689