Set ticket content to None and status to closed if linked sound does not exist
Becuase it can happen that some tickets have linked content which has dissapeared or on sound deletion time the ticket was not propertly updated, we need to check whether the sound that is linked does in fact exist. If it does not, we set the linked content to None and the status of the ticket to closed as should have been set at sound deletion time.

Further fixes in sound model need to be carried out too so that we should never find this situation.

https://github.com/MTG/freesound/issues/699