Use django auth urls and views instead of our customized ones
We now use django's views and urls for resetting password and for logout.
Login one we keep our one because of the extra info messages it provides.
Reset password views were overwritten in the past to check non-activated users, but in newer versions of Django this is already taken care.