Remove initial license error on multiple file describe
For change in the way Django handles initial data in forms, when loading the describe license page after selecting multiple sounds for description, an error appeared indicating the current license choice is not valid. Now no license is provided by default, forcing users to click on the one they prefer.