Create temporary license file for pack download
issue #751.
Removed create_license_file method from Pack and locations method
aswell. Now we create a text file the license information that is
temporary, this mean it has to be deleted in another process.
Also, the code to generate this zip file is generatic so it can
be used to generate a file with any set of sounds.