Verify that we can create a URL before using it
Fixes #824
We pass the `current_forum_name_slug` parameter from the view into
the template, which means that a user could change this value into
something not accepted by the URL string. Store the URL to a temp
value and check that it exists before using it. This may be a problem
in other views too.