Refactor after download modal to be better configurable
I renamed some things to remove references to donations in after download modal urls and views. The idea is that there is a view that is called after download and that decides whether to return any content or not. The after download modal can be configured to suggest donations, suggest participating in the survey or other stuff we might decide. This is selected in settings.py and content is defined in the after download modal view code.

https://github.com/MTG/freesound/pull/820