Add message warning for beta state of API monitoring
We do apply two different rate limits depending on whether the requests are for uploading content or for other types. We should probably have two charts, one per type of request. Otherwise it could be confusing if some users see that they are being limited but the chart does not reflect it. Also I'm not sure how will the chart display 429s. I guess these won't appear, but we should make them appear above the rate limit line so that users can see how much they are requesting. 429s are stored in the API errors stream, but I think are not assigned to a client id so what I'm proposing requires also work on improving logging on API errors like 429s. We should open new tickets for that, current functionality will probably be ok for 90% of uses.

https://github.com/MTG/freesound/pull/847