[WIP] Simplify logic for showing duplicate email help page
Some properties have been implemented in SameUser object to simplify the logic. Now what we do is that if a user is in the SameUser table, he is always redirected to the page for fixing duplicated email problems. Before loading this page, we check if the user did in fact fix the problems and if that is the case then we delete the SameUser entry and rediect the user. If that is not the case we show the message.

Simplification of the logic was possible due to the fact that we know there are no more than two accounts with shared emails.

This commit still needs proper testing.

https://github.com/MTG/freesound/pull/763