Add a probability to show the after download modal
Once all the requirements are met to decide that a modal should be shown, we still add some randomness to the decision by only showing it with a specific probability. This parameter works in combination with the maximum number of times to show a modal during a day. It means that the maximum will still be respected but instead of showing the modal at every download click, it will be shown only sometimes.

https://github.com/MTG/freesound/issues/916