Fix "An invalid form control with name='files'..." error
As described in https://stackoverflow.com/questions/22148080/an-invalid-form-control-with-name-is-not-focusable, this error happened because the browser was trying to do form validation (client-side) on a required field (`files` field) which was hidden. By setting `required=False` the issue gets fixed. This is not a problem in our case because our form does server-side validation on the number of uploaded files, therefore if no files are uploaded a message is shown to the user anyway.

Set `required=False` in