Disable links to donate via Stripe
Stripe account is not fully configured there fore we disabled the link and JS code to donate via Stripe. This is because the Stripe branch was already merged into master before the account was fully configured. Another option would be to add a new settings parameter to enable/disable donations via stripe but I considered this to be a temporal thing that does not require such setting.