Catch socket.error for connection refused when deleting sounds from solr
I introduced a bug in commit 7082fd97 where I removed an `except
Exception` and replaced it with SolrException. This except block was
causing tests to pass when running with an unconfigured solr server,
because it would catch the Connection Refused socket error.

Tests shouldn't rely on blanket exceptions to catch misconfigurations,
so I added mocks for all tests that delete sounds, to just check that
this method was called, instead of calling it.