Choose random sound solr parameter based on rand, not time
The random_x parameter in solr gives the same response for the same
value of x, so it must be chosen randomly. We originally decided to use
time() for this, because if we used randint we would have to make sure
that the maximum value was greater than the number of sounds in the
database. However, as x increases the returned sound id from solr
increases the same amount - random_1 -> 120, random_2 -> 122, etc,
so this doesn't feel very random.

Replace the time with a random number between 1 and 10 million
(which hopefully will contain the total number of sounds for
quite some while yet)