Use tvars in search.search view and util function to perform query
In order to enable good unit testing for the `search.search` view we moved some of its functionality into a `perform_solr_query` util function. In this way we'll be able to mock it. We took the opportunity to clean the code of the view, including the use of `tvars` to pass variables to the templates instead of `locals()`.