Remove unused `invalid` parameter
The `invalid` parameter was used to suppress the query from appearing in
the search box if "search!" was clicked without typing anything in the
box. In this case, no results were returned and the placeholder text
was not put in the search box.
Due to an error in the Watermark init code, it correctly works if the
input box is filled with the template text.

Also fix a longstanding bug with the position of ' in quoting the forum
name