Remove user profile fields in admin, add link to profile template
Adding the profile as inline to the user admin page was causing the page to never load for some unknown reason. Same thing was observed locally. This needs further investigation and considering the number of PRs that we have and that this feature is not a priority, I opted for the easy solution of just adding a link to the profile next to the link to the user.