Update ratings fixture to be more explicit about num/avg ratings
We have a fixture for sounds which includes sounds with num_ratings
and average_ratings but had no Ratings objects. Update the fixture to
make these values 0 so that the tests make sense with the number of
Ratings objects created in each test.

Add an explicit test for deleting all ratings for a sound and improve
the readability of the aggregate function used to compute the average