Add related names to Download, Post and Sound models
The related names are so that we can access downloads of a sound or pack with object.downloads, and to access posts of a user as user.posts and sounds of a pack as pack.sounds. This is needed so that the annotation Count(related_name) in report_count_statuses.py work properly. Django does not allow to use default related names (i.e. 'sounds_set') for this case.