Improve speed of report_count_statuses command
Speed is improved by using .annotate and .extra to reduce the number of needed DB queries.
This versions of the command also fixes a bug of the previous one, in which only uploaders where checked for the consistency of num_sounds and num_posts values and not all potential users (all uploaders and forum posters).