Admin improvements to handle username changes
Improve error message in admin when trying to set username to one that already exists. Also add search in OldUsername admin. This will be useful if a user contacts support because he want's to go back to a username that he used in the past. The corresponding OldUsername object will need to be deleted before renaming the user via the admin or the edit profile form.

https://github.com/MTG/freesound/pull/1088