Add warning note about `fields` param in API docs
This warning note should make more visible to API users reading the docs that the `fields` parameter can be used for specifying which metadata should be returned for each sound in a search request. This should help reducing the pattern of 1 search request to get N results and N extra to get the details for each result.