Re-add `pack` field in `Download` model and change related name
In https://github.com/MTG/freesound/pull/1114, we previously removed the `pack` field in `Download` model but left out the migration that made this change in the database. We decided that we want to keep this field in https://github.com/MTG/freesound/pull/1114 and only remove it in a later PR. In this commit we re-add the field.

When re-adding the field it happens that there is a related name collision in `Pack` for `Download` and `PacksDownload` tables. Therefore in this commit I rename the related name from `Pack` to `Download` as it will be removed anyway in https://github.com/MTG/freesound/pull/1115.