Only build sound map if user has geotagged sounds
We move code for building the sound map using the Google Maps API inside the "if nSounds > 0" condition, meaning that we will only mae a request to the maps API whe needed. Before that change we were always calling the maps API when loading a user profile (even if user does not have geotagged sounds).