Only build sound map if pack has geotagged sounds
We move code for building the sound map using the Google Maps API inside the "if nSounds > 0" condition, meaning that we will only mae a request to the maps API when needed. Before that change we were always calling the maps API when loading a pack page (even if the pack does not have geotagged sounds).