Add cache name back to cache tag.
I incorrectly thought that all parameters of the cache template tag
were variables to make the cache on, but the first parameter is a
name. This was causing users to see cached pages for other user's
pack lists. Bug introduced in db15d13e7c
reported at https://freesound.org/forum/bug-reports-errors-and-feature-requests/41145/