Add object previews in spam report emails
Now spam email reports sent to admin will include a small preview of the reported object so that most of the times that a decision about what to do with the user can be taken from the email client without the need to inspect each reported object individually.

This commit also includes some small refactorings to the user flagging code and adds unit tests.

https://github.com/MTG/freesound/issues/1218