Explicit warning now dismisses in moderation
I refactored the code so the dismiss button calls a function called
"remove_explicit_content_warning" and no post-binding of the "a" tag
is necessary. This fixes the problem in the moderation page in which the
sound players were loaded after the binding code for the dismiss buttons
was run. It also fixes potential similar future issues.

https://github.com/MTG/freesound/issues/1227