Process query filters to allow human-friendly AC field names
Because of the need of having different dynamicField in Solr for
different field types, we need to append a suffix to the indexed field
name that matches with the field type (bool, str, ...). Because we don't
want API users to be using these suffixes we use internally, we process
the filter contents to automatically transform "normal" AC field names
to the corresponding field name with the required suffix.

https://github.com/MTG/freesound/issues/1232