Replace uses of hardcoded content type ID in SoundManager
For some of the queries we do in SoundManager methods we need to get
Sound's content type ID. We use to hardcode this number (which made the
tests fail because testing DB has different content type IDs). Now
this number is obtained using the content types framework (no extra
queries to DB are made).

https://github.com/MTG/freesound/pull/1238