Remove <> from around URLs in emails that we send. Fixes #1257
We've seen a few places where web-based email clients tend to
eat text inside <> tags, which is where we traditionally put
URLs (#1257, #901) even though the email is marked as text/plain.

As the angle brackets don't bring too much to the message (they're
already all on their own lines) I propose that we remove the brackets.