Check if a username is valid when checking against existing users
Fixes #1258
We give feedback to users before they create their account if the
username is already taken, but if the username is invalid (e.g.
has a @ symbol in it) then we say that it's available, and then fail
the validation after the form is submitted.
Now we do a Validation of the field before checking if the username
exists to give faster feedback to the user.