Refactoring process/analyze functions into a class
This was to better handle storing out processing log to sound table in
db. Using a class we can keep on storing log messages as they are logged
and then store to db object. Using the classes code is also better
organized and I was able to de-duplicate code. Also, in this commit
there are a number of other minor refactorings:

- Removed essentia timeout as there is already a timeout in the worker
- Simplified logging code