Simplify analysis/processing cleanup of tmp files
Analysis and processing functions now use a context manager to create
a temporary folder. This folder is automatically deleted when the
function finished running (including if it finishes because of an
exception). This change allows to significantly simplify the cleaning
process and the creation of tmp folders which no longer need to be
passed as arguments from the worker.