Improve sending posts to solr
When we updated counts in threads and forums, we used F-expressions, and
then sent post objects to solr. This meant that we were setting
num_posts to values like "F(num_posts) + Value(1)" in solr, which is
incorrect.
Change these endpoints to take a post id, and perform a fresh get of the
objects before adding them to solr.

Optimise the Post query in order to avoid having to perform many queries
when doing bulk updates.

Update all posts when a thread is saved, so that the details about the
thread (title, num posts) are updated if these change

Fixes #1302