Change stripe integration from legacy to lastest version
Now we create a session with the donation info from the view,
then we pass the session_id to the frontend which makes a
redirect to stripe with the given session.

After the user completes the transaction it's redirected to the
'success' page which displays a notification.

Now we create the Donation object only when stripe sends us a
notification using webhooks.

issue #1307