Update test to avoid sound of the day cache problems
This test passed locally but not in travis. It looks like it is because
having some "sound of the day" stored in a cache but not present in DB.
We should further look into this, but because this is not anyway the
object of this test we changed the URL to make it pass.