Add cache.clear() in uni test to avoid error
Apparently the test "test_request_email_change" failed because of
requesting "front-page" view and having set a "random_sound" cache
key while no sounds being loaded in DB. Django seems to not clear
caches between tests so in this case we need to do it manually to make
the test pass in all circumstances.

Running tests locally did not get to this error, but this could be due
to tests being run in different order and producing different cache
effects.