Restore the behaviour of using a cached count for the paginator
We have a Paginator subclass which will take a count of the number of
items being paginated so that we don't have to call count() on a
queryset. We used an implementation detail of django to write to
Paginator._count, but this was changed in a recent version of django so
our paginators were calling qs.count() again.
Add a new count property that bypasses calling .count() if a count is
provided to the paginator.