Perform less queries when showing item flag links (#1344)
There was a lot of additional complexity in the flag_user template tag
that wasn't used. By performing a join in the view and passing in some
slightly different objects to the template tag, we can avoid running a
query per item (comment, message, post) in order to determine if we
should show a report link.
Some testing still required to determine if this is faster in all cases
(it's very slow in the 'all comments' view)