Only check profile fields in license and tos middlewares
All users who have profile.has_old_license set should have uploaded
sounds (pending some bugs reported in #1354), and so there's no reason
to also check the number of sounds too.

Because in 569495d3cce we load the profile on initial page load, it's
not an extra query to get this information. This means that there's no
reason to add extra complexity to this method with the cache, the
checking should be just as fast directly on the profile object each
request.