Cache latest sounds more aggressively
The DelayedQueryExecuter was being evaluated on the sounds page, which
meant that it wasn't serving its purpose. Replace with Manager.raw,
which also isn't evaluated until it's used. Fixes #1343.

Share the latest sound block between the front page and the sounds page
so that it is always loaded from cache when available.

Add an index to optimise the query to get the latest sounds so that it's
fast even when the cache isn't available.