Improve speed of most downloaded sounds (#1360)
We were evaluating this query in every view in order to use the
display_raw_sound template tag, causing the template cache to not have
an effect. Instead, use display_sound so that we can take advantage of
the cache (we do 5 lookups for tags, but these are small).
Also use greatest(created, moderation_date) in check condition to take
advantage of our indexes.
Query time has reduced from ~800ms to ~300ms, and then we cache this
value for an hour.

Also select related user information on pack downloads to reduce number
of queries