Refactor display_sound templatetag
The templatetag has been refactored to optimize the number of DB queries
performed depending on the argument passed to the templatetag. With this
refactored I removed the display_raw_sound templatetag and now we always
use display_sound. Also I added  a 'player_size' parameter and a new
'display_sound_big' templatetag (wrapper to display_sound) needed for
the BW interface.

https://github.com/MTG/freesound/pull/1364