Revert wrong fix for email_to and user_to assertion
In my previous commit (aa9bf11aa5d8172cce6f2fe32ea14b20269afe37) I submitted a
wrong fix for the assertion that made tests fail. The issue was not with the
assertion but with the fact that SUPPORT variable was empty by default in
Freesound Django settings and that would make the assertion fail when sending
an email to support. This commit adds a default email to SUPPORT to avoid
this issue.

https://github.com/MTG/freesound/issues/1427