Small refactor in tags views to facilitate future testing
While investigating for #1491 I came across this code and I made some small refactoring to facilitate testing of this method in the future. I in fact added a test for the view that used the refactoring (mocking solr call like for search views tests) but later I removed it in favour of directly testing the issue in the paginator. However I thought I'd leave this refactoring as it can be useful in the future.