Construct change event manually
This is needed because the way select elements react to changes is by adding an event listener for "onchange" event instead of by adding the property "onchange=". This is because we should not include references to JS functions from HTML tempaltes as the names of the functions will change once post-processed in the build chain. Therefore, all linking from DOM elements and JS functions must be done on the JS side with event listeners.