Submit main search form when pressing enter key
We need to do that by adding an event listener at the body element instead on only at the form element. This is because otherwise when changing checkboxes these do not get focus and the form element won't catch the keypress event. We add a check so that form is only submitted if search input is visible.