Only load user and pack maps when users specifically request it
This affects old interface and something similar needs to be ported to the new interface. We need to implement such a measure because mapbox pricing has changed and we are now being charged 8x times more. We might find a better solution but for now we need to reduce number of map loads.