WIP major search refactoring
Now using SearchEngine and SearchResults classes instead of the Solr wrappers as intermediaries. The goal for using this classes is to completely separate concepts specific to solr form the application side. In this way it will be easier to update and test different search engine backends (thinking of future solr versions for example)