Add search with solr 5.5.5
This adds a new solr core, with a new configuration and schema
(not copied from the v4.5 core), as many solr type classes were
removed between these versions.

This change also removes JTS in favour of solr's built-in geo filtering

Introduce a new solr backend for this version using pysolr instead of the
custom solr client.
This backend uses json to add documents rather than xml, but we still
keep the doc cleanups which remove control characters that are invalid
in xml.