Add solr9 freesound core
The new PointInteger field in solr doesn't support being used as
a unique id field, so we switched id to a string. This requires
casting it back to an int in a few places when using it to retrieve
sounds from the database

date fields can be a DatePoint or DateRange, the Range type is used
for filtering within a range (e.g. created:[from TO to]), but the
point field must be used for ordering, so we add the created field
with both types and choose one depending on what we want to do.