Unique tags before adding them to solr
Some sounds have the same tag multiple times, in upper and lower case.
This has the result of boosting sounds with multiple tag instances higher
in the search results (as solr sees that the term frequency is higher).
In order to prevent this unfair boost, unique and lower-case all tags
before adding to the index.
This won't prevent boosting from occurring if a word appears multiple times
in other search fields, such as description. (until we introduce the
BooleanSimilarity)