Add search tests for get_user_tags and get_pack_tags
With the change to solr 5, we didn't notice that the behaviour
of get_pack_tags changed from AND to OR, and so the list of
tags in a user's pack was incorrect, instead showing the most
common tags that the user had applied over all of fs, not just
within this pack. Explicitly use AND in the search query.
We're doing a query against /select and not using dismax, so use
set_query.