Rename old solr backend to common utils
The two classes that were remaining in solr451custom were common utils
for use in the solr backend, so rename the file to better reflect its
purpose instead of copying the definitions in to the solr backend.

Remove the custom SolrQueryPySolr class, as this was only needed to
separate it from the custom backend, which is no longer used.
Rename the as_dict method to as_kwargs to better reflect its purpose.