Refactor sound display caches
Now, in BW, sound display caches don't depend on is_autjenticated and is_explicit variables and thesea re not used in the cache key. We achieve that by making bookmark button always visible (even for logged out users), and by leaving the "is explicit" part out of the cached html.