Don't convert with ffmpeg if file is uncompressed
This reverts a change I did some weeks ago in which I added ffmpeg conversion to PCM even for wav/aiff files. This was done because some files would fail processing otherwise and a conversion was later triggered anyway. However, this causes problems because it changes the bitdepth of the original files and then "stereofy and find info" gets wrong bitdepth values. In the future we could fix that by separating the "stereofy" from the "get info" process, and get info from the original files instead of the transcoded versions.