Move the geotag relation from sound.geotag to geotag.sound
Only a very small proportion of sounds have a geotag, and there can only
be one geotag per sound. Therefore it makes more sense to put a sound
reference on the geotag instead of a geotag on the sound model which is
almost always null. This should reduce the size of the sounds table