set USE_TZ=True
With postgresql, this doesn't cause any issues with setting or
retrieving data, even if the server and database server are set to
different timezones.
In order to not compare offset-naive/aware timezones, always use
django's timezone.now() method to get now.
Make sure timezone aware dates are used in all fixtures and migrations