use the ORM for bulk_query methods
We originally wrote this query in 2015 when django couldn't do
subqueries. This ability was added in 2017!
We did some testing and using the ORM is very slightly faster than our
custom code, but more or less the same.
One major change is to get the analyzers data in a prefetch_related
instead of a subquery. Despite the additional query, the time needed to
get the data is quite similar