Write packed sounds bytes in a single operation
This was set up to write a single byte at a time, as sound id, lat, lng
However, if a lat or lng value was an invalid number, ValueError would
get raised. In this case, we would have written only one (id) or two
(id, lat) bytes to the list before continuing.
By writing all 3 bytes at once, if any of the conversions fails the
entire sound will be skipped.