Add workflow for indexing similarity vectors to solr
Adding vectors to solr is slow (due to us needing to read them from
disk), so we prefer to do this in two steps, first adding regular
text fields then another pass which updates the docs adding the vectors.

We had an option to add only specific fields, but this was a generic
solution which wasn't used anywhere, remove it because it was
overly-complex.

Add a flag to reindex_search_engine_sounds --include-similarity-vectors
to include similarity vectors in the initial collection creation
Add a flag --only-similarity-vectors to update the existing solr
collection and add only similarity vectors.

Modify post_dirty_sounds_to_search_engine to always perform a document
update so that it doesn't clear the associated child documents which
have the vectors