Always recreate forum index, add logging, correctly get post ids
post ids were being returned as strings from solr  so the code to
reindex was deleting them because they didn't match the int from the
database. Cast to int.
However, we also remove this code because we now create a new index on
demand - we should never be in a position where we try and reindex onto
an existing index