Add a new string field for faceting tags
In solr 9.7, some changes were made to the way that faceting works
(SOLR-12963). This means that to facet on a field, it must have
docValues=true set, which means it must be a string.

Our tag field is still "text", which is good for searching but cannot
have docValues=true set, so make a new field and copy tags to it. It
doesn't need to be stored.

In the tag browse page, facet on this new field