Don't allow multiple bookmarks to be created with no category
We had a unique constraint on (user, sound, category) so we can't put
the same sound more than once in a category.
However if category is null then postgres doesn't count these as
"the same", so it's possible to add the same sound more than once under
"no category". Add a new constraint and clean up these dup bookmarks
before applying it.