When removing a category, delete bookmarks which would become a conflict
We have a unique constraint on (user, sound_id) when category is null.
When we delete a category we keep the bookmarks but remove the category.
If there is already a bookmark for this sound with no category then we
get a unique constraint error when trying to set a category to null.
Delete any bookmarks from this category that would cause a problem
before deleting the category.