Only rewrite duration filter if it's a valid solr range
We rewrite a filter of the form "duration:[lower TO upper]", but we get
some queries of "duration:lower TO upper" which parses as a string not a
range.

In the case that the duration filter is invalid report an error to the
user and don't continue with the search.