Handle race conditions in pack edit, tag update, similarity vectors
If we get two rapid requests in succession (for example if a user
double-clicks on a button instead of single click) then we start two
transactions and they can both believe that changes need to be made,
causing an integrity error when adding a unique item for the second
time.

Use get_or_create which will skip making an item if it already exists.
In pack edit, catch IntegrityError on save and report this error to the
user