Use get_or_create in Collections to prevent creating duplicate items
If we get two rapid requests in succession (for example if a user
double-clicks on a button instead of single click) then we start two
transactions and they can both believe that changes need to be made,
causing an integrity error when adding a unique item for the second
time. Use get_or_create which does this in a transaction and will
correctly return an existing item.