Add ticket messages directly in POST instead of in background task
If a ticket is returned the assigned moderator is removed. This value
was used in the background task to set the author of the message, but
for a returned ticket this is now None, so the author is set to
anonymous.  Add the message immediately while we still know the author,
this operation isn't very expensive so doesn't need to be deferred to
the task.